+++ lib/arachni/support/database/base.rb
@@ -6,8 +6,6 @@
     web site for more information on licensing and terms of use.
 =end
 
-require 'tmpdir'
-
 module Arachni
 module Support::Database
 
@@ -96,7 +94,7 @@ class Base
 
     def generate_filename
         # Should be unique enough...
-        "#{Arachni.tmpdir}/#{self.class.name}_#{Process.pid}_#{object_id}_#{@filename_counter}".gsub( '::', '_' )
+        "#{Options.paths.tmpdir}/#{self.class.name}_#{Process.pid}_#{object_id}_#{@filename_counter}".gsub( '::', '_' )
     ensure
         @filename_counter += 1
     end