+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+## Under development
+
 ## 1.5 _(January 31, 2017)_
 
 - Executables